requirements.txt
authy==2.2.2
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
Flask-Login==0.4.1
futures==3.2.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-core==0.29.1
googleapis-common-protos==1.5.8
grpcio==1.18.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn>=19.6.0
google-cloud-datastore>=1.7.3
google-endpoints==4.8.0
google-endpoints-api-management==1.11.0

